Rename unauthorised handler service
